// src/communityRepository/communityMembership/observers/searchMembers/enums.ts
export enum EnumCommunityMemberActions {
  OnCommunityJoined = 'onCommunityJoined',
  OnCommunityLeft = 'onCommunityLeft',
  OnCommunityUserBanned = 'onCommunityUserBanned',
  OnCommunityUserChanged = 'onCommunityUserChanged',
  OnCommunityUserRoleAdded = 'onCommunityUserRoleAdded',
  OnCommunityUserRoleRemoved = 'onCommunityUserRoleRemoved',
  OnCommunityUserUnbanned = 'onCommunityUserUnbanned',
}